+++ processor/elastictraceprocessor/processor.go
@@ -19,18 +19,15 @@ package elastictraceprocessor // import "github.com/elastic/opentelemetry-collec
 
 import (
 	"context"
+	"github.com/elastic/opentelemetry-collector-components/processor/elastictraceprocessor/internal/ecs"
+	"github.com/elastic/opentelemetry-collector-components/processor/elastictraceprocessor/internal/index"
 	"github.com/elastic/opentelemetry-lib/enrichments/trace"
 	"go.opentelemetry.io/collector/client"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/collector/processor"
-	semconv26 "go.opentelemetry.io/collector/semconv/v1.26.0"
-	semconv "go.opentelemetry.io/collector/semconv/v1.27.0"
 	"go.uber.org/zap"
-	"strings"
-	"time"
 )
 
 var _ processor.Traces = (*Processor)(nil)
@@ -57,38 +54,12 @@ func (p *Processor) Capabilities() consumer.Capabilities {
 }
 
 func (p *Processor) ConsumeTraces(ctx context.Context, td ptrace.Traces) error {
-
 	resourceSpans := td.ResourceSpans()
 	for i := 0; i < resourceSpans.Len(); i++ {
 		resourceSpan := resourceSpans.At(i)
 		resource := resourceSpan.Resource()
-		translateResourceMetadata(resource)
-		attributes := resource.Attributes()
-
-		serviceName, _ := attributes.Get("service.name")
-
-		dataStreamType := "traces"
-
-		attributes.PutStr("data_stream.type", dataStreamType)
-		attributes.PutStr("data_stream.dataset", "apm."+serviceName.AsString())
-		attributes.PutStr("data_stream.namespace", "default") // Use a default or configurable namespace
-
-		scopeSpans := resourceSpan.ScopeSpans()
-		for j := 0; j < scopeSpans.Len(); j++ {
-
-			scopeSpan := scopeSpans.At(j)
-			spans := scopeSpan.Spans()
-
-			for k := 0; k < spans.Len(); k++ {
-
-				span := spans.At(k)
-				attributes := span.Attributes()
-				timestamp := span.StartTimestamp().AsTime().Format(time.RFC3339Nano)
-				attributes.PutStr("timestamp", timestamp)
-
-			}
-
-		}
+		ecs.TranslateResourceMetadata(resource)
+		index.EncodeDataStream(resource, "traces")
 	}
 	p.enricher.Enrich(td)
 	ecsCtx := client.NewContext(ctx, withMappingMode(client.FromContext(ctx), "ecs"))
@@ -101,106 +72,4 @@ func withMappingMode(info client.Info, mode string) client.Info {
 		Auth:     info.Auth,
 		Metadata: client.NewMetadata(map[string][]string{"x-elastic-mapping-mode": {mode}}),
 	}
-}
-
-func translateResourceMetadata(resource pcommon.Resource) {
-	attributes := resource.Attributes()
-
-	attributes.Range(func(k string, v pcommon.Value) bool {
-		if !isSupportedAttribute(k) {
-			attributes.PutStr("labels."+replaceDots(k), v.AsString())
-			attributes.Remove(k)
-		}
-		return true
-	})
-}
-
-func replaceDots(key string) string {
-	return strings.ReplaceAll(key, ".", "_")
-}
-
-func isSupportedAttribute(attr string) bool {
-	switch attr {
-	// service.*
-	case semconv.AttributeServiceName,
-		semconv.AttributeServiceVersion,
-		semconv.AttributeServiceInstanceID,
-		semconv.AttributeServiceNamespace:
-		return true
-
-	// deployment.*
-	case semconv26.AttributeDeploymentEnvironment, semconv.AttributeDeploymentEnvironmentName:
-		return true
-
-	// telemetry.sdk.*
-	case semconv.AttributeTelemetrySDKName,
-		semconv.AttributeTelemetrySDKVersion,
-		semconv.AttributeTelemetrySDKLanguage:
-		return true
-
-	// cloud.*
-	case semconv.AttributeCloudProvider,
-		semconv.AttributeCloudAccountID,
-		semconv.AttributeCloudRegion,
-		semconv.AttributeCloudAvailabilityZone,
-		semconv.AttributeCloudPlatform:
-		return true
-
-	// container.*
-	case semconv.AttributeContainerName,
-		semconv.AttributeContainerID,
-		semconv.AttributeContainerImageName,
-		"container.image.tag",
-		"container.runtime":
-		return true
-
-	// k8s.*
-	case semconv.AttributeK8SNamespaceName,
-		semconv.AttributeK8SNodeName,
-		semconv.AttributeK8SPodName,
-		semconv.AttributeK8SPodUID:
-		return true
-
-	// host.*
-	case semconv.AttributeHostName,
-		semconv.AttributeHostID,
-		semconv.AttributeHostType,
-		"host.arch",
-		semconv.AttributeHostIP:
-		return true
-
-	// process.*
-	case semconv.AttributeProcessPID,
-		semconv.AttributeProcessCommandLine,
-		semconv.AttributeProcessExecutablePath,
-		"process.runtime.name",
-		"process.runtime.version",
-		semconv.AttributeProcessOwner:
-		return true
-
-	// os.*
-	case semconv.AttributeOSType,
-		semconv.AttributeOSDescription,
-		semconv.AttributeOSName,
-		semconv.AttributeOSVersion:
-		return true
-
-	// device.*
-	case semconv.AttributeDeviceID,
-		semconv.AttributeDeviceModelIdentifier,
-		semconv.AttributeDeviceModelName,
-		"device.manufacturer":
-		return true
-
-	// data_stream.*
-	case "data_stream.dataset",
-		"data_stream.namespace":
-		return true
-
-	// Legacy OpenCensus attributes
-	case "opencensus.exporterversion":
-		return true
-	}
-
-	return false
 }
